Make paginator context-aware to use buttons in slash command context
